v2.4.0 Add get_children_async (#25)
In order to fetch children for devices that support them, specifically HS300 and KP303, the sys_info for the "parent" device has to be requested. This extra call can add up when there are multiple parent devices associated with your account and the total time spent to get devices would grow linearly with the number of devices you have with children. This can be especially painful when using the "prefetch" feature for constructing a device manager as you won't be able to use the manager until all of those requests have been completed.

With this change, children can be gathered async and so rather than wait on each device's `get_children` method sequentially, this will instead gather up all of the async "children gathering" tasks, then wait on the lot of them to complete.

This change was made with the intent of being backwards-compatible so as to avoid breaking any uses of "public" methods (those without a `_` preceding their name).

Note that the original `get_children` method has been moved to the `TPLinkDevice` class as it simply calls the new async version.