v2.6.0 Add has_emeter convenience function
This function is intended to be an easy way to check if a device has emeter capabilities and therefore can get electricity data.

Note that this also includes a bugfix - HS300 devices themselves as an entity here do not have emeter functionality. It is their children (HS300Child - the individual plugs on the power strip) that have that functionality.