v2.6.1 Fix uncaught exception on Windows
This change is a Windows OS-specific HACK to silence an exception thrown on the asyncio event loop being closed as part of the asyncio library's proactor.

Note that the exception is technically harmless to the library's functionality, but it is a poor user experience to see exceptions with normal usage.

Hack sourced from an issue on the aiohttp library: https://github.com/aio-libs/aiohttp/issues/4324#issuecomment-733884349