v2.7.0 Add "Power Tools"
This change adds a new importable class, TPLinkDeviceManagerPowerTools, which builds upon the previous TPLinkDeviceManager adding functionality specifically pertaining to emeter devices. The main benefit of this toolset is that requests are handled to all matching devices asynchronously, so if you are wanting to gather power usage data across a large number of devices, it will occur very quickly. This is a convenience class and does not change the core functionality of the device manager.

There are however some changes to the TPLinkEMeterDevice, which adds new async versions of the existing functionality alongside them.

The bulk of this change is in test files. I do not like the mostly-duplicate content of the request mappings but decided not to dig further into alternatives at this time.