v2.9.0 Add Rule Building
This change introduces a new class, TPLinkDeviceScheduleRuleBuilder to complement added methods exposing device functionality around rules for a device schedule. Rules can not be retrieved by their ID, new rules can be added, and existing rules can be deleted.

The existing functionality exposed to edit a rule is now more user friendly with the TPLinkDeviceScheduleRuleBuilder class which allows a rule to be built or modified . Examples of this are provided in the README.md