v3.0.0 Separate Login Functionality
The `TPLinkDeviceManager` can now be constructed without a username and password. In order for requests to work properly, either the new `login` function or `set_auth_token` will need to be called.

While this change should not realistically break any existing functionality, an error is no longer raised when providing a `None`-valued username or password to the `TPLinkDeviceManager` constructor meaning that code depending on that behavior _would_ be broken.

The intent behind this change is to allow a device manager to be constructed that is "light weight", meaning it is a simple class instance that will not do anything like login when constructed. This separation allows a shell instance to be constructed without knowledge of a username and password, and then an auth token to be set before making requests. This bypasses the need to "login" for uses of the manager that may only have an auth token at the time instead of a username and password.

What this allows for is a workflow wherein a user of the manager can explicitly login, and cache or store the resulting auth token for later use. Later, the auth token could be retrieved and set in the manager for requests, without needing the original username and password.